Game Intermission - Platformer Part #1 - Getting Started

+++ GAMEENGINE/ENGINE/ECS/COMPONENTS/AnimationComponent.h
@@ -7,7 +7,7 @@
 struct AnimationComponent
 {
 	int numFrames{ 1 }, frameRate{ 1 }, frameOffset{ 0 }, currentFrame{ 0 };
-	int startTime{ static_cast<int>(glfwGetTime()) };
+	double startTime{ glfwGetTime() };
 	bool bVertical{ false }, bLooped{ false };
 	static void CreateAnimationLuaBind(sol::state& lua);
 };
